apps/job-hunter/pyproject.toml
[project]
name = "ironclaw-job-hunter"
version = "0.1.0"
description = "AI Job Hunter - Automated job application system using MobileRun AI"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.76.0",
    "flask>=3.1.2",
    "google-api-python-client>=2.188.0",
    "google-auth>=2.47.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.2.4",
    "openai>=2.15.0",
    "pydantic>=2.12.5",
    "pymongo>=4.16.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "httpx>=0.28.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/job_hunter"]

[project.scripts]
job-hunter = "job_hunter.main:main"
